Split main.cpp into multiple classes
