+++ CMakeLists.txt
@@ -3,4 +3,10 @@ project(Task_Manager)
 
 set(CMAKE_CXX_STANDARD 17)
 
-add_executable(Task_Manager main.cpp)
+add_executable(Task_Manager src/main.cpp
+        src/Util.cpp
+        src/Util.h
+        src/Program.cpp
+        src/Program.h
+        src/Task.cpp
+        src/Task.h)
