+++ src/Task.h
@@ -1,0 +1,21 @@
+//
+// Created by herob on 2023-11-23.
+//
+
+#ifndef TASK_MANAGER_TASK_H
+#define TASK_MANAGER_TASK_H
+#include <string>
+
+using namespace std;
+
+
+class Task {
+public:
+    string name;
+    string date;
+    string description;
+    string priority;
+};
+
+
+#endif //TASK_MANAGER_TASK_H
